Ditch IE/Edge support
Formerly, IE and Edge worked with a Custom Elements polyfill.
Now, IE is flat-out unsupported (you’d need to polyfill its
XMLHttpRequest so that it could handle data: URIs and implement
responseType 'arraybuffer', as well as TextEncoder), and Edge would
probably work with a TextEncoder polyfill, but I’m not supporting it.

No more Array; now everything is ArrayBuffer.